Extract util module for React filters
